feat: print address information registered by user
- 현재 사용자가 등록한 주소를 보기 위해서 rest api를 통해 외부
ap현재 지도 검색 결과시 위도, 경도 좌표 위치가 반대로 되어있어
이를 올바르게 수정
- 통일성을 위해서 js 를 jsx형태로 고치고 파일명 형식을 현재
프로젝트의 스타일을 따름(main)
- 외부 api를 호출하는 경우는 별도의 service를 두어 하나로 합침
