+++ web-apps/client/src/pages/Area.jsx
@@ -1,4 +1,4 @@
-import React, { useContext, useCallback } from 'react';
+import React, { useContext } from 'react';
 
 import MapAdjust from '../components/MapAdjust';
