+++ web-apps/client/src/pages/Main.jsx
@@ -1,5 +1,5 @@
-import React, {useState, useEffect, useContext} from 'react';
-import {Link} from 'react-router-dom';
+import React, { useState, useEffect, useContext } from 'react';
+import { Link } from 'react-router-dom';
 
 import {
   Typography,
@@ -7,26 +7,26 @@ import {
   GridListTile,
   CircularProgress,
 } from '@material-ui/core';
-import {makeStyles} from '@material-ui/core/styles';
+import { makeStyles } from '@material-ui/core/styles';
 import SearchIcon from '@material-ui/icons/Search';
 import FilterIcon from '@material-ui/icons/Tune';
 import ClearIcon from '@material-ui/icons/Clear';
 
-import ActionBar from '../../components/ActionBar';
-import ProductListItem from '../../components/ProductListItem';
+import ActionBar from '../components/ActionBar';
+import ProductListItem from '../components/ProductListItem';
 
-import getButtons from '../../utils/action-bar';
+import getButtons from '../utils/action-bar';
 
-import { getProductList } from './fetch';
-import { filterContext } from '../../contexts/Filters';
-import { AlertMessageContext } from '../../contexts/AlertMessage';
-import { SnackbarContext } from '../../contexts/SnackBar';
+import { getProductList } from '../service/product';
+import { filterContext } from '../contexts/Filters';
+import { AlertMessageContext } from '../contexts/AlertMessage';
+import { SnackbarContext } from '../contexts/SnackBar';
 
-import { ROUTES } from '../../assets/uris';
+import { ROUTES } from '../assets/uris';
 
-import {isScrollBottom, debounce} from '../../utils';
-import useScrollDown from '../../hooks/useScrollDown';
-import useWindowResize from '../../hooks/useWindowResize';
+import { isScrollBottom, debounce } from '../utils';
+import useScrollDown from '../hooks/useScrollDown';
+import useWindowResize from '../hooks/useWindowResize';
 
 const useStyles = makeStyles({
   list: {
@@ -52,7 +52,7 @@ const Main = () => {
   const [loading, setLoading] = useState(false);
   const [cols, setCols] = useState(1);
   const [list, setList] = useState([]);
-  const [settings, setSettings] = useState({from: 0, limits: 10});
+  const [settings, setSettings] = useState({ from: 0, limits: 10 });
 
   const clearKeyword = (event) => {
     event.preventDefault();
@@ -73,7 +73,7 @@ const Main = () => {
       }
       setLoading(true);
       try {
-        const result = await getProductList({...filter, ...settings});
+        const result = await getProductList({ ...filter, ...settings });
         setList((state) => [...state, ...result]);
       } catch (e) {
         dispatchMessage({ type: ACTION_TYPE.ERROR, payload: '데이터를 불러오는 중 오류가 발생하였습니다.' });
@@ -116,13 +116,13 @@ const Main = () => {
               {name}
             </Typography>
           </Link>
-)}
+        )}
         title={headerTitle}
         buttons={buttons}
       />
       <GridList spacing={0} cols={cols} className={classes.list}>
         {list.map(
-          ({id, hits, title, pictures, price, order, distance, interests}) => (
+          ({ id, hits, title, pictures, price, order, distance, interests }) => (
             <GridListTile key={id} className={classes.list}>
               <Link to={`/product/${id}`}>
                 <ProductListItem
