--- web-apps/client/src/utils/config.js
@@ -1,2 +1,0 @@
-export const PRODUCT_API = 'http://localhost:5000';
-export const KAKAO_API = '0e7e14bdabdac7f7f61b7093185b8996';
